refa: 상속을 통한 중복 코드 제거
