+++ src/main/java/player/Participant.java
@@ -1,0 +1,42 @@
+package player;
+
+import java.util.List;
+
+import card.Card;
+
+abstract class Participant {
+    protected final Name name;
+    protected final Hand hand;
+
+    public Participant(Name name, Hand hand) {
+        this.name = name;
+        this.hand = hand;
+    }
+
+    public void hit(Card card) {
+        hand.add(card);
+    }
+
+    public int calculateScore() {
+        return hand.calculateScore();
+    }
+
+    public List<Card> showCards() {
+        return hand.getCards();
+    }
+
+    public boolean isBust() {
+        return hand.calculateScore() >= 22;
+    }
+
+    public Name getName() {
+        return name;
+    }
+
+    abstract void win();
+
+    abstract void lose();
+
+    abstract void tie();
+
+}
