+++ src/main/java/player/Player.java
@@ -1,37 +1,12 @@
 package player;
 
-import java.util.List;
-
 import blackjackGame.Result;
-import card.Card;
 
-public class Player {
-    private final Name name;
-    private final Hand hand = new Hand();
+public class Player extends Participant {
     private Result result;
 
     public Player(Name name) {
-        this.name = name;
-    }
-
-    public void hit(Card card) {
-        hand.add(card);
-    }
-
-    public int calculateScore() {
-        return hand.calculateScore();
-    }
-
-    public List<Card> showCards() {
-        return hand.getCards();
-    }
-
-    public boolean isBust() {
-        return hand.calculateScore() >= 22;
-    }
-
-    public Name getName() {
-        return name;
+        super(name, new Hand());
     }
 
     public void win() {
